move sql query split to sqlutil
