+++ devtools/psql-lite/main.go
@@ -2,10 +2,13 @@ package main
 
 import (
 	"context"
+	"errors"
 	"flag"
+	"fmt"
 	"log"
 
-	"github.com/target/goalert/test/smoke/harness"
+	"github.com/jackc/pgx/v5"
+	"github.com/target/goalert/util/sqlutil"
 )
 
 func main() {
@@ -23,13 +26,47 @@ func main() {
 		defer cancel()
 	}
 
-	var err error
-	if *inTx {
-		err = harness.ExecSQLBatch(ctx, *db, *cmd)
-	} else {
-		err = harness.ExecSQL(ctx, *db, *cmd)
-	}
+	conn, err := pgx.Connect(ctx, *db)
 	if err != nil {
 		log.Fatal(err)
+	}
+	defer conn.Close(ctx)
+
+	var tx pgx.Tx
+	if *inTx {
+		tx, err = conn.Begin(ctx)
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer sqlutil.RollbackContext(ctx, "psql-lite: exec sql", tx)
+	}
+
+	for _, q := range sqlutil.SplitQuery(*cmd) {
+		rows, err := conn.Query(ctx, q)
+		if errors.Is(err, pgx.ErrNoRows) {
+			break
+		}
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		for rows.Next() {
+			var s string
+			err = rows.Scan(&s)
+			if err != nil {
+				log.Fatal(err)
+			}
+			fmt.Println(s)
+		}
+		if err := rows.Err(); err != nil {
+			log.Fatal(err)
+		}
+	}
+
+	if *inTx {
+		err = tx.Commit(ctx)
+		if err != nil {
+			log.Fatal(err)
+		}
 	}
 }
